refactor(content-update): factor out content functions into helpers file
